requirements.txt
# requirements.txt

PyGithub==2.5.0
PyYAML==6.0.2
python-dotenv==1.0.1
jinja2==3.1.6
reportlab==3.6.13